Changed border, started saveBoard
saveBoard will drop a compressed image of the board into a text box.  Select and copy it to a text file to save.
I'm having some trouble accessing the  HTML table that stores the board.  I got the idea of putting an id on each tinput box as I generate it.  See how that goes.